Generacion de terrenos mejorada
Se cambiar los valores porcentuales por rangos de tamano. Se tiene que revisar como se crearan los terrenos ( Cada cuanto)

+++ terrainPrototype.h
@@ -13,6 +13,6 @@ public:
 	int gripRange[2];
 
 	TerrainPrototype(std::string pName,int * ranges);
-	Terrain * getTerrain(int currentDistance, int finalDistance, float minPercentag, float maxPercentage);
+	Terrain * getTerrain(int currentDistance, int finalDistance, int minPercentag, int maxPercentage);
 	std::string toString();
 };
